Rename ssl test files to match Scrutinice filter
