+++ ssl/CMakeLists.txt
@@ -83,8 +83,8 @@ if(BUILD_TESTING)
     ssl_test.cc
     ssl_c_test.c
     ssl_key_share_test.cc
-    ssl_alps_tests.cc
-    ssl_test_common.cc
+    ssl_alps_test.cc
+    ssl_common_test.cc
     ssl_misc_test.cc
     ssl_version_test.cc
     ssl_ech_test.cc
