+++ ssl/ssl_quic_test.cc
@@ -3,7 +3,7 @@
 
 #include <openssl/ssl.h>
 #include "../crypto/test/test_util.h"
-#include "ssl_test_common.h"
+#include "ssl_common_test.h"
 
 BSSL_NAMESPACE_BEGIN
